[psg-20210202/#expand-new-api-with-tests] -- working tests for: Nonowner can not delete my vaultfolder; Can not delete my root vaultfolder

+++ app/Http/Controllers/VaultfoldersController.php
@@ -157,6 +157,9 @@ class VaultfoldersController extends AppBaseController
         if ( $request->user()->cannot('delete', $vaultfolder) ) {
             abort(403);
         }
+        if ( $vaultfolder->isRootFolder() ) {
+            abort(400);
+        }
         $vaultfolder->delete();
         return response()->json([]);
     }
